Missed a size_t

+++ apc_sma.c
@@ -397,7 +397,7 @@ void apc_sma_free(void* p)
     LOCK(sma_lock);
 
     for (i = 0; i < sma_numseg; i++) {
-        size_t d_size = (unsigned int)((char *)p - (char *)(sma_shmaddrs[i]));
+        size_t d_size = (size_t)((char *)p - (char *)(sma_shmaddrs[i]));
         if (p >= sma_shmaddrs[i] && d_size < sma_segsize) {
             sma_deallocate(sma_shmaddrs[i], d_size);
             if (APCG(mem_size_ptr) != NULL) { *(APCG(mem_size_ptr)) -= d_size; }
